discord-bot-template/requirements.txt
py-cord==2.4.1
feedparser==6.0.11
requests==2.31.0
aiohttp==3.9.5
python-dotenv==1.0.1
